+++ src/main/java/com/aaax/web/AdminPortalController.java
@@ -10,4 +10,19 @@ public class AdminPortalController {
     public String admin() {
         return "forward:/admin/index.html";
     }
+
+    @GetMapping({"/sign-in", "/sign-in/"})
+    public String signIn() {
+        return "forward:/sign-in/index.html";
+    }
+
+    @GetMapping({"/sign-up", "/sign-up/"})
+    public String signUp() {
+        return "forward:/sign-up/index.html";
+    }
+
+    @GetMapping({"/user", "/user/"})
+    public String user() {
+        return "forward:/user/index.html";
+    }
 }